backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "This is backend for Project Inpersona"
authors = [
    {name = "yatharth",email = "yatharthkapadia2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "llama-index-core (==0.11.19)",
    "llama-index-embeddings-huggingface (==0.3.1)",
    "llama-index-llms-groq (==0.2.0)",
    "llama-index-vector-stores-chroma (==0.2.1)",
    "ipython (>=8.32.0,<9.0.0)",
    "pyvis (>=0.3.2,<0.4.0)",
    "chromadb (==0.5.15)",
    "colorama (>=0.4.6,<0.5.0)",
    "transformers (==4.45.2)",
    "llama-index-readers-file (==0.2.2)",
    "llama-index-llms-gemini"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
